+++ pkg/orkestra-registry/pvs/pv.go
@@ -140,7 +140,7 @@ func Resolve(src orktypes.PVTemplateSource, ownerName string) ResolvedPVSpec {
 	for _, l := range src.Labels {
 		spec.Labels[l.Key] = l.Value
 	}
-	spec.Labels[labels.Managed] = labels.ManagedValue
+	spec.Labels[labels.ManagedKey] = labels.ManagedValue
 	spec.Labels[labels.OrkestraOwner] = ownerName
 
 	return spec