+++ pkg/orkestra-registry/replicasets/replicaset.go
@@ -229,7 +229,7 @@ func Resolve(src orktypes.ReplicaSetTemplateSource, ownerName string) ResolvedRe
 
 	spec.Env = []orktypes.EnvVar(src.Env)
 
-	spec.Labels[labels.Managed] = labels.ManagedValue
+	spec.Labels[labels.ManagedKey] = labels.ManagedValue
 	spec.Labels[labels.OrkestraOwner] = ownerName
 
 	return spec